Refactor of the arduino protocol
The firmware code has been converted into a Arduino Library. The new design simplifies the interface and create the abstraction for manage different types of connection (In theory, any type of serial interface that extends from Stream class should work well).

Also, the new design allows the final user to add new code to be executed in the board (this features can affect the protocol performance, so should be used with caution).

Issue #22